agentcore-browser-agent/requirements.txt
bedrock-agentcore>=0.1.3
strands-agents>=1.7.1
strands-agents-tools>=0.2.7
playwright>=1.40.0
nest-asyncio>=1.5.0
boto3>=1.39.7
botocore>=1.39.7
typing-extensions>=4.0.0